refactor refresh button
